chore: factor out ContainerEntry and QList::Entry into a common class
